# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A job board API for managing job postings",
        "title": "Job Board API",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/jobs": {
            "get": {
                "description": "Search for jobs with optional filters and pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Search for jobs",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"golang developer\"",
                        "description": "Search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "example": 20,
                        "description": "Number of results to return (max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "example": 0,
                        "description": "Number of results to skip",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Experience level filter",
                        "name": "experience",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Employment type filter",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "costarica",
                            "latam"
                        ],
                        "type": "string",
                        "example": "\"costarica\"",
                        "description": "Location filter",
                        "name": "location",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "remote",
                            "hybrid",
                            "onsite"
                        ],
                        "type": "string",
                        "example": "\"remote\"",
                        "description": "Work mode filter",
                        "name": "mode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"Tech Corp\"",
                        "description": "Company name filter (partial match)",
                        "name": "company",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"2024-01-01\"",
                        "description": "Start date filter (YYYY-MM-DD)",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"2024-12-31\"",
                        "description": "End date filter (YYYY-MM-DD)",
                        "name": "date_to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/jobs.SearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/jobs.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jobs.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/jobs/count": {
            "get": {
                "description": "Get the total count of jobs that match the search criteria without returning the actual job data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Get count of jobs matching search criteria",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"golang developer\"",
                        "description": "Search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Experience level filter",
                        "name": "experience",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Employment type filter",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "costarica",
                            "latam"
                        ],
                        "type": "string",
                        "example": "\"costarica\"",
                        "description": "Location filter",
                        "name": "location",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "remote",
                            "hybrid",
                            "onsite"
                        ],
                        "type": "string",
                        "example": "\"remote\"",
                        "description": "Work mode filter",
                        "name": "mode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"Tech Corp\"",
                        "description": "Company name filter (partial match)",
                        "name": "company",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"2024-01-01\"",
                        "description": "Start date filter (YYYY-MM-DD)",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"2024-12-31\"",
                        "description": "End date filter (YYYY-MM-DD)",
                        "name": "date_to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns count in format: {\"count\": 42}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/jobs.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jobs.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "jobs.ErrorDetails": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "jobs.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/jobs.ErrorDetails"
                }
            }
        },
        "jobs.JobRequirementsResponse": {
            "type": "object",
            "properties": {
                "must_have": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "nice_to_have": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "jobs.JobResponse": {
            "type": "object",
            "properties": {
                "application_url": {
                    "type": "string"
                },
                "benefits": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "company_id": {
                    "type": "integer"
                },
                "company_logo_url": {
                    "type": "string"
                },
                "company_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "employment_type": {
                    "type": "string"
                },
                "experience_level": {
                    "type": "string"
                },
                "job_id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "posted_at": {
                    "type": "string"
                },
                "requirements": {
                    "$ref": "#/definitions/jobs.JobRequirementsResponse"
                },
                "responsibilities": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "technologies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/jobs.TechnologyResponse"
                    }
                },
                "title": {
                    "type": "string"
                },
                "work_mode": {
                    "type": "string"
                }
            }
        },
        "jobs.PaginationDetails": {
            "type": "object",
            "properties": {
                "has_more": {
                    "type": "boolean"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "jobs.SearchResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/jobs.JobResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/jobs.PaginationDetails"
                }
            }
        },
        "jobs.TechnologyResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "required": {
                    "type": "boolean"
                }
            }
        }
    }
}